+++ packages/sdk/src/evm/functions/getContract.ts
@@ -2,11 +2,11 @@ import { resolveAddress } from "../common/ens";
 import { PREBUILT_CONTRACTS_MAP } from "../contracts";
 import { SmartContract } from "../contracts/smart-contract";
 import { ContractPublisher } from "../core/classes/contract-publisher";
+import { NetworkInput } from "../core/types";
 import {
   ContractForPrebuiltContractType,
-  NetworkInput,
   PrebuiltContractType,
-} from "../core/types";
+} from "../contracts";
 import { AddressOrEns } from "../schema/shared";
 import { SDKOptions } from "../schema/sdk-options";
 import { getContractFromAbi } from "./getContractFromAbi";
