+++ packages/sdk/src/evm/functions/utils/cache.ts
@@ -1,4 +1,4 @@
-import { ValidContractInstance } from "../../core/types";
+import { ValidContractInstance } from "../../contracts";
 import { ThirdwebStorage } from "@thirdweb-dev/storage";
 
 let STORAGE_CACHE = new ThirdwebStorage();
